+++ src/gh_profiler/utils/profile_utils.py
@@ -151,11 +151,31 @@ def _parse_pr_activity(pr_activity_str):
     prs = search["nodes"]
 
     pdata.opened_count = len(prs)
-    pdata.merged_count = sum(pr["mergedAt"] is not None for pr in prs)
-    pdata.closed_count = sum(
-        pr["state"] == "CLOSED" and pr["mergedAt"] is None for pr in prs
+
+    # PRs against repos the user owns.
+    prs_owned = [
+        pr for pr in prs
+        if pr["repository"]["owner"]["login"].casefold()
+        == pdata.username.casefold()
+    ]
+
+    pdata.opened_count_owned = len(prs_owned)
+    pdata.merged_count_owned = sum(pr["mergedAt"] is not None for pr in prs_owned)
+    pdata.closed_count_owned = sum(
+        pr["state"] == "CLOSED" and pr["mergedAt"] is None for pr in prs_owned
     )
 
+    # PRs against external repos.
+    prs_external = [
+        pr for pr in prs
+        if pr["repository"]["owner"]["login"].casefold()
+        != pdata.username.casefold()
+    ]
+    pdata.opened_count_external = len(prs_external)
+    pdata.merged_count_external = sum(pr["mergedAt"] is not None for pr in prs_external)
+    pdata.closed_count_external = sum(
+        pr["state"] == "CLOSED" and pr["mergedAt"] is None for pr in prs_external
+    )
 
 def _fetch_issue_activity():
     """Fetch target user's recent public issue activity."""
@@ -211,21 +231,27 @@ def _get_pr_query():
             search(query: $q, type: ISSUE, first: $n) {{
                 issueCount
                 pageInfo {{
-                hasNextPage
-                endCursor
+                    hasNextPage
+                    endCursor
                 }}
                 nodes {{
-                ... on PullRequest {{
-                    number
-                    state
-                    createdAt
-                    closedAt
-                    mergedAt
-                    url
-                }}
+                    ... on PullRequest {{
+                        number
+                        state
+                        createdAt
+                        closedAt
+                        mergedAt
+                        url
+                        repository {{
+                            nameWithOwner
+                            owner {{
+                                login
+                            }}
+                        }}
+                    }}
                 }}
             }}
-            }}
+        }}
     """
 
     return dedent(pr_query).strip()